Changes Dev to Staging in CI workflow and scripts
Previously, the instance in which the containers were deployed to and
AMI built from was called the Dev Server. Given that it's more of a
'pre-production' environment, I've changed the name to Staging Server.
This commit changes all variables and CI files that refer to the Dev
Server to the Staging Server.