Adds step to RunContainers.yml to stop Containers
Previously, the compose file was copied to the staging server and then
the older containers were stopped and images removed. This posed a
problem where the new compose file could use a different tag for the
containers, and then the old containers would not be removed. This
commit fixes that issue.