Adds DeregisterAmi workflow
Adds a workflow to be triggered manually that will deregister AMI and
delete snapshot associated with it. Either the 'oldest' or 'latest'
version of the Lyria AMI can be chosen. Additionally, adds workflow
to push the current AMI to production, but nothing is actually in that
file as of now.