Adds manual workflow (PushAmiToProduction)
The new workflow can be triggered manually and will push the latest AMI
to production in the environment selected. As of now, the workflow has
been tested locally up until the manual approval stage. Once the
workflow is verified to work remotely, steps to run terraform apply,
commit & push changes, create pull request, and merge the terraform
coniguration with the new AMI variable need to be added.