Changes IPv6 CIDR for Docker backend network
Changes IPv6 CIDR for backend network in an attempt to fix issue of
App container not responding to requests from the host. Additionally,
removes the 'protocol' arguments from the ssh-action in the
RunContainers.yml workflow, in an attempt to allow SSH to work with the
new staging server using an IPv6 address.