Configures blue environment for IPv6
This commit will configure the blue environment to use a single instance
with IPv6 and no public IPv4 addresses. In addition to using a single
instance and no IPv4, the current AMI does not use containers. Some more
troubleshooting will need to be done to use with Docker containers.
The current configuration works but the following resources still need
to be added:

- CloudFront Distribution for the main site
- SSL certificates for CloudFront Distribution
- Route53 DNS records for the CloudFront Distribution