Adds the new blue environment sans IPv4 addresses
New configuration works by placing ELB and Autoscaling group in private
subnets. Access to ELB DNS is achieved through creating a VPC link and
an API in API Gateway.

Next Steps:
CloudFront needs to be added to the confuration using the provisioned
API as an origin. Also need to configure SSL certificates, then modify
the DNS records to point to CloudFront distribution.