Changes Terraform config to blue environment
This commit successfully modifies the configuration to not use any
public IPv4 addresses. It utilizes an internal load balancer routing to
an autoscaling group, both in private subnets. A VPC link is used for
ALB connection to an API, which also routes to a CloudFront
distribution. The CloudFront distribution has been added to the Route53
DNS records (A and AAAA). Confirmed to work and green environment has
been destroyed.