Corrects the previous DNS issue infrastructure
DNS now mostly works. In chrome meettheafflerbaughs.com re-routes to
https://www.meettheafflerbaughs.com and it works perfectly. However, in
safari, simply typing in meettheafflerbaughs.com is not working. The
full https with www needs to be typed in. Furthermore, simply typing
https://meettheafflerbaughs.com does not work in either browser. This
will need to be troubleshooted further, but for now, all links to the
site include the full path with www, so there should be no issue
temporarily.