Begins migration to new AWS account
Creates storage buckets in new AWS account to being the migration of
Lyria to the new AWS account. Unlike the last account, there will now
be a separate bucket for dev and prod. Buckets are currently provisioned
but the specific CloudFront distribution will need to be added to the
bucket policy, as well as a CORS policy.