Removes a -y from the user data field in terraform
This commit removes a -y from one of the commands in the user data field
for the staging instance Terraform configuration. The addition of the -y
was preventing Docker from being installed due to the fact that -y was
an invalid option for that command. Additionally, this commit adds the
production bucket to the instance profile policy. Both prod and dev
buckets will be needed in both the production and staging environments
since the difference in dev and prod buckets is to test new songs,
rather than app features.