Configured blue environment for use in new account
This commit configures the blue environment for use in the new account.
At this point everything has been confirmed to work, except I am waiting
for the SSL certificate to be validated. As of now, the site is
accessible throught the CloudFront URL, but not through the domain. Once
the certificate is validated the certificate will need to be attached to
the CloudFront distribution and the ELB. Then records will need to be
updated in Route53.