Fixes issue with api returning 400 statuses
Fixes the issue with the API Gateway returning 400 statuses, which for
some reason was trying to send plain text http request to the https load
balancer listener. By simply removing the https listener and making the
API send http to the http listener, the issue was resolved. Since there
is already a certificate on the CloudFront distribution, and there is no
sensitive information being transmitted on the site, this is a safe
alternative.