Adds Lambda authorizer to API gateway
This commit bring the current configuration in line with the current
configuration and adds the Lambda authorizer to API Gateway, which was
provisioned outside of Terraform. There is still a problem with the
instances getting overloaded, terminating, and the autoscaling group
provisioning extra instances. Hoping this commit will update the AMI,
maybe it's just a problem with patching needing to be updated.