Adds configuration files for DynamoDB Table
This new table will contain the song order. Future commits will have the
app get the song order from the table rather than the hardcoded list,
which will not force the creation of a new AMI if song order changes.

The new commit adds the Terraform configuration for the table, as well
as a script in the hooks directory to change the song order.