Refactors to get song order from DynamoDB table
This commit refactors the index function to retrieve the song order from
the DynamoDB table. Also, it adds logic to change bucket key of audio
and image files based on which bucket is used (dev or prod), since the
dev bucket uses the 'dev' prefix. Debug is also automatically configured
based on whether or not the dev or prod bucket is used.